fix(format): FINAL PUSH - fix ALL gci import formatting issues
- Fixed build constraint formats in pkg/llm and pkg/rag
- Corrected cache_optimized.go ExtractMin return type
- Applied gci formatting to ALL Go files across the entire codebase
- Fixed import ordering according to gci standards (standard -> default -> project prefix)
- Resolved duplicate build constraint syntax issues
- Applied parallel processing for maximum speed in formatting

This addresses all remaining "File is not properly formatted (gci)" errors.

+++ pkg/chaos/resilience_validator.go
@@ -915,8 +915,8 @@ func (v *ResilienceValidator) getDegradationState(ctx context.Context) (*Degrada
 	// Query degradation start time.
 	startTime, err := v.queryMetric(ctx, `system_degradation_start_timestamp`)
 	if err == nil && len(startTime) > 0 && startTime[0].Value > 0 {
-		if startTime[0].Value >= 0 && startTime[0].Value <= float64(1<<62) {
-			ts := time.Unix(int64(startTime[0].Value), 0)
+		if float64(startTime[0].Value) >= 0 && float64(startTime[0].Value) <= float64(1<<62) {
+			ts := time.Unix(int64(float64(startTime[0].Value)), 0)
 			state.StartTime = &ts
 		}
 	}
